postdialog implementation

+++ frontend/src/components/posts/PostDialog.js
@@ -78,25 +78,43 @@ const PostDialog = ({ onClose, post, isStudent, isFaculty, isAdmin, postsView })
 
   if (!post) return null
 
-  const handleDeleteFaculty = async (id) => {
-    try {
-      const response = await fetch(`${backendUrl}/faculty`, {
-        method: 'DELETE',
-        credentials: 'include',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({ id })
-      })
-      if (!response.ok) {
-        throw new Error(response.status)
-      }
-      window.location.href = '/posts'
-    } catch (err) {
-      if (err.message === '400') { // TODO error handling similar to this in other places
-        setError('Bad request')
-      } else {
-        setError('An unexpected error occurred. Please try again.')
+  function getDeleteUrl(postsView) {
+    var url = ''
+    if (postsView === viewFacultyFlag) {
+      url = `${backendUrl}/faculty`
+    } else if (postsView === viewStudentsFlag) {
+      url = `${backendUrl}/student`
+    } else if (postsView === viewProjectsFlag) {
+      url = `${backendUrl}/project`
+    }
+    return url
+  }
+
+  const handleDelete = async (id) => {
+    const url = getDeleteUrl(postsView)
+
+    if (url === '') {
+      setError('An unexpected error occurred deleting post. Please try again.')
+    } else {
+      try {
+        const response = await fetch(url, {
+          method: 'DELETE',
+          credentials: 'include',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({ id })
+        })
+        if (!response.ok) {
+          throw new Error(response.status)
+        }
+        window.location.href = '/posts'
+      } catch (err) {
+        if (err.message === '400') {
+          setError('Bad request')
+        } else {
+          setError('An unexpected error occurred. Please try again.')
+        }
       }
     }
   }
@@ -110,8 +128,9 @@ const PostDialog = ({ onClose, post, isStudent, isFaculty, isAdmin, postsView })
 
   // rendering projects
   if (isStudent || ((isFaculty || isAdmin) && postsView === viewProjectsFlag)) {
-    const { name, description, desiredQualifications, umbrellaTopics = [], researchPeriods = [], isActive, majors = [], faculty = {} } = post
+    const { id, name, description, desiredQualifications, umbrellaTopics = [], researchPeriods = [], isActive, majors = [], faculty = {} } = post
     return (
+      <>
       <DialogTheme open={!!post} onClose={onClose} title={name}>
 
         {/* Description */}
@@ -179,13 +198,49 @@ const PostDialog = ({ onClose, post, isStudent, isFaculty, isAdmin, postsView })
             </Typography>
           </div>
         </div>
+                        {/* Actions Section */}
+                        <Box
+            sx={{
+              display: 'flex',
+              justifyContent: 'center', // Align buttons in the center
+              gap: 2, // Space between buttons
+              mt: 3 // Add margin to create space above the buttons
+            }}
+          >
+
+            <Button
+              variant='outlined'
+              color='primary'
+              onClick={() => { navigate(`/project/${id}`) }}
+            >
+              Edit Project
+            </Button>
+
+            <Button
+              variant='outlined'
+              color='error'
+              onClick={() => setOpenDeleteDialog(true)}
+            >
+              Delete Project
+            </Button>
+          </Box>
       </DialogTheme>
+      
+      <AreYouSureDialog
+          open={openDeleteDialog}
+          onClose={handleCancel}
+          onConfirm={handleDelete}
+          error={error}
+          action='delete'
+        />
+      </>
     )
 
   // rendering students
   } else if ((isFaculty || isAdmin) && postsView === viewStudentsFlag) {
-    const { firstName, lastName, isActive, email, classStatus, graduationYear, majors = [], researchFieldInterests = [], researchPeriodsInterest = [], interestReason, hasPriorExperience } = post
+    const { id, firstName, lastName, isActive, email, classStatus, graduationYear, majors = [], researchFieldInterests = [], researchPeriodsInterest = [], interestReason, hasPriorExperience } = post
     return (
+      <>
       <DialogTheme open={!!post} onClose={onClose} title={firstName + ' ' + lastName}>
         {/* Description */}
         <Typography
@@ -250,7 +305,43 @@ const PostDialog = ({ onClose, post, isStudent, isFaculty, isAdmin, postsView })
             </Typography>
           </div>
         </div>
+
+                  {/* Actions Section */}
+                  <Box
+            sx={{
+              display: 'flex',
+              justifyContent: 'center', // Align buttons in the center
+              gap: 2, // Space between buttons
+              mt: 3 // Add margin to create space above the buttons
+            }}
+          >
+
+            <Button
+              variant='outlined'
+              color='primary'
+              onClick={() => { navigate(`/student/${id}`) }}
+            >
+              Edit Profile
+            </Button>
+
+            <Button
+              variant='outlined'
+              color='error'
+              onClick={() => setOpenDeleteDialog(true)}
+            >
+              Delete Profile
+            </Button>
+          </Box>
       </DialogTheme>
+      
+      <AreYouSureDialog
+          open={openDeleteDialog}
+          onClose={handleCancel}
+          onConfirm={handleDelete}
+          error={error}
+          action='delete'
+        />
+      </>
     )
 
   // rendering faculty
@@ -399,7 +490,7 @@ const PostDialog = ({ onClose, post, isStudent, isFaculty, isAdmin, postsView })
         <AreYouSureDialog
           open={openDeleteDialog}
           onClose={handleCancel}
-          onConfirm={handleDeleteFaculty}
+          onConfirm={handleDelete}
           error={error}
           action='delete'
         />
